--- LegalDD/documents/runserver.bat
@@ -1,1 +1,0 @@
-manage.py runserver
